Fix feature specs after adding submission request
